Revert "Menubar"
