+++ GSB-RV-DR/nbproject/project.properties
@@ -78,7 +78,7 @@ javafx.enable.concurrent.external.runs=false
 javafx.enabled=true
 javafx.fallback.class=com.javafx.main.NoJavaFXFallback
 # Main class for JavaFX
-javafx.main.class=fr.gsb.rv.dr.Appli
+javafx.main.class=gsb.rv.dr.GSBRVDR
 javafx.preloader.class=
 # This project does not use Preloader
 javafx.preloader.enabled=false
